+++ lib/puppet/provider/package/cisco.rb
@@ -17,10 +17,11 @@
 require 'puppet/util/package'
 
 Puppet::Type.type(:package).provide :cisco, parent: :yum do
-  desc "The cisco nexus package provider.
+  desc "The cisco package provider.
   Local rpm installations will utilize the native yum provider.
   Cisco rpm installations from the host will utilize the native yum provider.
-  Cisco rpm installations from guestshell will utilize nxapi to install to host."
+  Cisco NX-OS rpm installations from guestshell will use this provider to install to host.
+  Cisco IOS XR rpm installations will utilize this provider to install to host via sdr_instcmd."
 
   confine feature: :cisco_node_utils
 
@@ -30,7 +31,7 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
   # these commands must exist to execute native yum provider
   commands yum: 'yum', rpm: 'rpm', python: 'python'
 
-  defaultfor operatingsystem: :nexus
+  defaultfor operatingsystem: [:ios_xr, :nexus]
 
   # if the following commands aren't present, we're in trouble
   if command('rpm')
@@ -45,6 +46,42 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
     confine true: commands_present
   end
 
+  # Returns true if operating system is nexus else false.
+  def in_nexus?
+    os = Facter.value('operatingsystem')
+    (os == 'nexus') ? true : false
+  end
+
+  # Returns true if operating system is ios_xr else false.
+  def in_ios_xr?
+    os = Facter.value('operatingsystem')
+    (os == 'ios_xr') ? true : false
+  end
+
+  # Returns true if rpm is ios-xr rpm. otherwise false.
+  def cisco_rpm_xr?
+    name_var_arch_regex_xr = /^(.*\d.*)-([\d.]*)-(r\d+.*)\.(\w{4,}).rpm/
+    if @resource[:source]
+      @resource[:source].match(name_var_arch_regex_xr) ? true : false
+    elsif @resource[:name] && @resource[:package_settings] && @resource[:platform]
+      source = @resource[:name] + '-' + @resource[:package_settings]['version'] + \
+               '.' + @resource[:platform] + '.rpm'
+      source.match(name_var_arch_regex_xr) ? true : false
+    end
+  end
+
+  # Returns either ios xr version or rpm package version, based upon input.
+  def version?(t)
+    return unless @resource[:package_settings]
+    regex = /^([\d.]*)-(r\d+.*)$/
+    @resource[:package_settings]['version'].match(regex)
+    if t == 'package'
+      Regexp.last_match(1)
+    elsif t == 'xr'
+      Regexp.last_match(2)
+    end
+  end
+
   # IMPORTANT: it's useless to override self.instances and prefetch,
   # because we can't know whether to retrieve packages for native or GS
   # because target->host is specified on a per-package basis. Instead,
@@ -53,11 +90,18 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
   # this method in package.rb determines how to set ensure
   # @property_hash is empty at this point
   def properties
-    if in_guestshell? && target_host?
+    if (in_ios_xr?) ||
+       (in_guestshell? && target_host?)
       normalize_resource
+    else
+      super
+    end
+
+    if (in_ios_xr? && cisco_rpm_xr?) ||
+       (in_guestshell? && target_host?)
 
       is_ver = current_version
-      should_ver = @resource[:package_settings]['version']
+      should_ver = version?('package')
 
       # set absent if no version is installed, or if installed version
       # does not match @resource version (if should_ver is provided)
@@ -69,6 +113,18 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
 
       debug "determined package #{@resource[:name]} is #{status}"
       @property_hash = { ensure: status, version: is_ver }
+
+    elsif in_ios_xr?
+      info = super
+      if info[:version]
+        status = :present
+        ver = info[:version]
+      else
+        status = :absent
+        ver = ''
+      end
+      debug "determined package #{@resource[:name]} is #{status}."
+      @property_hash = { ensure: status, version: ver }
     else
       super
     end
@@ -91,7 +147,7 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
     # ex: b+z-ip2.x64_64
     name_arch_regex = /^([\w\-\+]+)\.(\w+)$/
 
-    if @resource[:name] =~ name_arch_regex
+    if in_nexus? && @resource[:name] =~ name_arch_regex
       @resource[:name] = Regexp.last_match(1)
       @resource[:platform] = Regexp.last_match(2)
       debug "parsed name:#{Regexp.last_match(1)}, arch:#{Regexp.last_match(2)}"
@@ -99,15 +155,35 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
     # [source] overrides [name]
     return unless @resource[:source]
 
+    # ex xrv9k-k9sec-1.0.0.0-r600.x86_64.rpm-6.0.0
+    #    xrv9k-k9sec-1.0.0.0-r61102I.x86_64.rpm-XR-DEV-16.02.22C
+    # This regex is used to create a match group of
+    # 1. package name, 2. package version, 3. platform
+    name_var_arch_regex_xr = /^(.*\d.*)-([\d.]*-r\d+.*)\.(\w{4,}).rpm/
+    name_regex_xr = /(.*).rpm$/
+
     # convert to linux-style path before parsing filename
     filename = @resource[:source].strip.tr(':', '/').split('/').last
 
-    if filename =~ name_ver_arch_regex ||
-       filename =~ name_var_arch_regex_nx
+    if in_ios_xr?
+      if filename =~ name_var_arch_regex_xr
+        @resource[:name] = Regexp.last_match(1)
+        @resource[:package_settings]['version'] = Regexp.last_match(2)
+        @resource[:platform] = Regexp.last_match(3)
+        debug "parsed name:#{Regexp.last_match(1)}," \
+          "version:#{Regexp.last_match(2)}, " \
+          "arch:#{Regexp.last_match(3)}"
+      else
+        @resource[:name] = filename.match(name_regex_xr)[1]
+        return
+      end
+    elsif filename =~ name_ver_arch_regex ||
+          filename =~ name_var_arch_regex_nx
       @resource[:name] = Regexp.last_match(1)
       @resource[:package_settings]['version'] = Regexp.last_match(2)
       @resource[:platform] = Regexp.last_match(3)
-      debug "parsed name:#{Regexp.last_match(1)}, version:#{Regexp.last_match(2)}, arch:#{Regexp.last_match(3)}"
+      debug "parsed name:#{Regexp.last_match(1)}," \
+          "version:#{Regexp.last_match(2)}, arch:#{Regexp.last_match(3)}"
     else
       @resource.fail 'Could not parse name|version|arch from source: ' \
         "#{@resource[:source]}"
@@ -118,7 +194,8 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
 
   # helper to retrieve version info for installed package
   def current_version
-    if @resource[:platform]
+    # Platform is not used in XR query.
+    if @resource[:platform] && in_nexus?
       ver = Cisco::Yum.query("#{@resource[:name]}.#{@resource[:platform]}")
     else
       ver = Cisco::Yum.query("#{@resource[:name]}")
@@ -159,9 +236,19 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
   end
 
   def install
-    if in_guestshell? && target_host?
-      debug 'Guestshell + target=>host detected, using nxapi for install'
+    if in_ios_xr?
+      if cisco_rpm_xr?
+        debug 'using sdr_instcmd for install'
+        Cisco::Yum.install("#{@resource[:name]}-" \
+                           "#{@resource[:package_settings]['version']}")
+      elsif @resource[:source]
+        debug "using yum for install #{@resource[:source]}"
+        @resource[:name] = @resource[:source]
+        super
+      end
+    elsif in_nexus? && (in_guestshell? && target_host?)
       if @resource[:source]
+        debug 'Guestshell + target=>host detected, using nxapi for install'
         Cisco::Yum.install(@resource[:source])
       else
         Cisco::Yum.install(@resource[:name])
@@ -177,15 +264,23 @@ Puppet::Type.type(:package).provide :cisco, parent: :yum do
   # yum's update method calls self.install which will refer to this class' install
 
   def uninstall
-    if in_guestshell? && target_host?
-      debug 'Guestshell + target=>host detected, using nxapi for uninstall'
+    if in_ios_xr?
+      if cisco_rpm_xr?
+        debug 'using sdr_instcmd for uninstall'
+        Cisco::Yum.remove("#{@resource[:name]}-"\
+                          "#{@resource[:package_settings]['version']}")
+      else
+        debug "using yum for uninstall #{@resource[:name]}"
+        super
+      end
+    elsif in_nexus? && (in_guestshell? && target_host?)
       if @resource[:platform]
         Cisco::Yum.remove("#{@resource[:name]}.#{@resource[:platform]}")
       else
         Cisco::Yum.remove(@resource[:name])
       end
     else
-      debug 'Not Guestshell + target=>host, use native yum provider for uninstall'
+      debug 'Not XR || Guestshell + target=>host, use native yum provider for uninstall'
       super
     end
   end